vhost_user: Move enum_value macro from message.rs to mod.rs
Move the enum_value macro into vhost_user module and make it public within
this module. This will be used in the next commit for defining gpu message
enum.
